Add performance module with three coarse regression tests
